-reenable film number as otherwise film selection in various tables does not work anymore

+++ src/main/java/mediathek/daten/DatenDownload.java
@@ -118,7 +118,7 @@ public final class DatenDownload implements Comparable<DatenDownload> {
         this.film = film;
         this.pSet = pSet;
         this.abo = abo;
-        arr[DOWNLOAD_FILM_NR] = Integer.toString(film.getFilmNr());
+        arr[DOWNLOAD_FILM_NR] = Long.toString(film.getFilmNr());
         arr[DOWNLOAD_SENDER] = film.getSender();
         arr[DOWNLOAD_THEMA] = film.getThema();
         arr[DOWNLOAD_TITEL] = film.getTitle();
